10/04
- added texture loading with stb_image
- in pixel shader, added optimization where in the raymarcher it first roughly checks the volume texture using a single lookup and then, if it is close enough, checks again using trilinear filtering (8 lookups)
- added compute shader (find_brush_cs.hlsl) that will return where the mouse is pointing and its normal
- refactored buffers, now buffers have both SRV and UAV and can be structured. binding and unbinding is decoupled to the buffer itself and can be done on multiple at one time.
- refactored shaders, now they only have 2 pointers (the shader itself and an "extra", either input layout or sample state) instead of 5. shaders can only represent ONE type of shader at a time
- refactored dynamic shader, now it acts like a collection of shaders that are updated
- fixed warning in ini.cc about double -> float conversion
- add fill function to arr<T> class
- added TODO file to keep track of what i need to do

+++ src/ini.cc
@@ -86,7 +86,7 @@ namespace ini {
 
 	void Value::trySet(float &val) const {
 		if (isValid()) {
-			val = asNum();
+			val = (float)asNum();
 		}
 	}
